+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @fileverse-dev/fortune-react
 
+## 1.1.66
+
+### Patch Changes
+
+- Fix bug with mobile scrolling
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.1.66
+
+## 1.1.65
+
+### Patch Changes
+
+- Improve mobile scroll UX and add indication for format menu
+- Updated dependencies
+  - @fileverse-dev/fortune-core@1.1.65
+
 ## 1.1.64
 
 ### Patch Changes